Updated AdminController to use our new Global strings and added checks to ensure endpoints validate users.

+++ InterCareBackend/Controllers/AdminController.cs
@@ -1,4 +1,5 @@
 ﻿using InterCareBackend.Daos.Implementations;
+using InterCareBackend.Helpers;
 using InterCareBackend.Models;
 using Microsoft.AspNetCore.Mvc;
 using System;
@@ -6,17 +7,28 @@ using System.Collections.Generic;
 
 namespace InterCareBackend.Controllers
 {
-    public class AdminController
+    public class AdminController : ControllerBase
     {
         OrganizationAdminDao adminDao = new OrganizationAdminDao();
         OrganizationDao organizationDao = new OrganizationDao();
+        AuthHelper auth = new AuthHelper();
 
 
         [HttpPost("/api/createAdminWithOrganization")]
-        public void create()
+        public String create()
         {
-            adminDao.createAdmin("Jonar17@student.sdu.dk", "Jonas Støve Rasmussen", "Pass", "0", "Admin");
-            organizationDao.createOrganization("Jonas's Organization", new List<String> { }, adminDao.getAdminByEmail("Jonar17@student.sdu.dk").id);
+            var header = Request.Headers["Authorization"].ToString().Substring("Bearer ".Length).Trim();
+            IDictionary<string, object> token = this.auth.decodeJWT(header);
+            if(token["type"].ToString() == Globals.GlobalInterCareAdmin)
+            {
+                adminDao.createAdmin("Jonar17@student.sdu.dk", "Jonas Støve Rasmussen", "Pass", "0", Globals.GlobalOrganizationAdmin);
+                organizationDao.createOrganization("Jonas's Organization", new List<String> { }, adminDao.getAdminByEmail("Jonar17@student.sdu.dk").id);
+                return Globals.GlobalValidType;
+            } else
+            {
+                return Globals.GlobalInvalidType;
+            }
+           
         }
 
 
@@ -24,14 +36,31 @@ namespace InterCareBackend.Controllers
 
         public OrganizationAdmin getAdminByEmail()
         {
-            return adminDao.getAdminByEmail("Jonar17@student.sdu.dk");
+            var header = Request.Headers["Authorization"].ToString().Substring("Bearer ".Length).Trim();
+            IDictionary<string, object> token = this.auth.decodeJWT(header);
+            if (token["type"].ToString() == Globals.GlobalInterCareAdmin)
+            {
+                return adminDao.getAdminByEmail("Jonar17@student.sdu.dk");
+            } else
+            {
+                return null;
+            }
         }
 
 
         [HttpDelete("/api/deleteAdminByEmail")]
-        public void delete()
+        public string delete()
         {
-            adminDao.deleteAdmin("Jonar17@student.sdu.dk");
+            var header = Request.Headers["Authorization"].ToString().Substring("Bearer ".Length).Trim();
+            IDictionary<string, object> token = this.auth.decodeJWT(header);
+            if (token["type"].ToString() == Globals.GlobalInterCareAdmin)
+            {
+                adminDao.deleteAdmin("Jonar17@student.sdu.dk");
+                return Globals.GlobalValidType;
+            } else
+            {
+                return Globals.GlobalInvalidType;
+            }
         }
